refactor(lotto): 코드 스타일 개선
